ml-service/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
pillow
transformers
torch
easyocr
